+++ lib/apiScenario/apiScenarioTypes.ts
@@ -136,6 +136,8 @@ export type StepRestCall = StepBase & {
   parameters: SwaggerExample["parameters"];
   responses: SwaggerExample["responses"];
   outputVariables?: OutputVariables;
+  externalReference?: boolean;
+  _resolvedParameters?: SwaggerExample["parameters"];
 };
 
 //#endregion
@@ -306,32 +308,31 @@ export type ScenarioDefinition = TransformRaw<
     scenarios: Scenario[];
     cleanUpSteps: Step[];
     _filePath: string;
+    _swaggerFilePaths: string[];
   }
 >;
 //#endregion
 
 //#region Runner specific types
-export interface RawReport {
-  executions: RawExecution[];
+export interface NewmanReport {
+  executions: NewmanExecution[];
   timings: any;
   variables: { [variableName: string]: Variable };
-  testScenarioName?: string;
-  metadata: any;
 }
 
-export interface RawExecution {
-  request: RawRequest;
-  response: RawResponse;
+export interface NewmanExecution {
+  request: NewmanRequest;
+  response: NewmanResponse;
   annotation?: any;
 }
-export interface RawRequest {
+export interface NewmanRequest {
   url: string;
   method: string;
   headers: { [key: string]: any };
   body: string;
 }
 
-export interface RawResponse {
+export interface NewmanResponse {
   statusCode: number;
   headers: { [key: string]: any };
   body: string;