+++ lib/swaggerValidator/trafficValidator.ts
@@ -16,7 +16,7 @@ import { traverseSwagger } from "../transform/traverseSwagger";
 import { Operation, Path, LowerHttpMethods } from "../swagger/swaggerTypes";
 import { LiveValidatorLoader } from "../liveValidation/liveValidatorLoader";
 import { inversifyGetContainer, inversifyGetInstance } from "../inversifyUtils";
-import { findPathsToKey, findPathToValue, getApiVersionFromSwaggerPath } from "../util/utils";
+import { findPathsToKey, findPathToValue, getApiVersionFromFilePath } from "../util/utils";
 import { SwaggerLoader, SwaggerLoaderOption } from "../swagger/swaggerLoader";
 import { getFilePositionFromJsonPath } from "../util/jsonUtils";
 
@@ -359,7 +359,7 @@ export class TrafficValidator {
 
       this.operationCoverageResult.push({
         spec: key,
-        apiVersion: getApiVersionFromSwaggerPath(key),
+        apiVersion: getApiVersionFromFilePath(key),
         coveredOperaions: coveredOperaions,
         coverageRate: coverageRate,
         unCoveredOperations: value.length - coveredOperaions,