+++ test/apiScenario/markdownReportTest.ts
@@ -6,19 +6,19 @@ import {
   generateMarkdownReport,
   generateMarkdownReportHeader,
 } from "../../lib/apiScenario/markdownReport";
-import { TestScenarioResult } from "../../lib/apiScenario/reportGenerator";
+import { ApiScenarioTestResult } from "../../lib/apiScenario/newmanReportValidator";
 
 describe("markdownReport", () => {
   it("Should generate markdown report", () => {
     const ts = {
-      testScenarioFilePath: "Microsoft.Compute/preview/2020-09-30/test-scenarios/galleries.yaml",
+      apiScenarioFilePath: "Microsoft.Compute/preview/2020-09-30/test-scenarios/galleries.yaml",
       swaggerFilePaths: ["Microsoft.Compute/preview/2020-09-30/gallery.json"],
       providerNamespace: "Microsoft.Compute",
       apiVersion: "2020-09-30",
       runId: "202106011456-d4udg",
       rootPath: "/home/zhenglai/repos/azure-rest-api-specs/specification/compute/resource-manager",
       environment: "test",
-      testScenarioName: "galleries_1",
+      apiScenarioName: "galleries_1",
       armEndpoint: "https://management.azure.com",
       stepResult: [
         {
@@ -56,7 +56,7 @@ describe("markdownReport", () => {
       startTime: "2021-06-01T06:56:46.835Z",
       endTime: "2021-06-01T06:58:07.066Z",
       subscriptionId: "db5eb68e-73e2-4fa8-b18a-46cd1be4cce5",
-    } as TestScenarioResult;
+    } as ApiScenarioTestResult;
     const header = generateMarkdownReportHeader();
     const body = generateMarkdownReport(ts);
     expect(header).toMatchSnapshot("header");