+++ test/modelValidatorTests.ts
@@ -818,4 +818,17 @@ describe("Model Validation", () => {
       assert.strictEqual(result[0].code, "XMS_EXAMPLE_NOTFOUND_ERROR");
     }, 10000);
   });
+
+  describe("format validation", () => {
+    it("should fail when value is not in base64 format", async () => {
+      const specPath2 = `${testPath}/modelValidation/swaggers/specification/formatValidation/format.json`;
+      const result = await validate.validateExamples(specPath2, "Byte");
+      assert.strictEqual(result.length, 1);
+      assert.strictEqual(result[0].code, "INVALID_FORMAT");
+      assert.strictEqual(
+        result[0].message,
+        "Object didn't pass validation for format byte: credentialValue1"
+      );
+    });
+  });
 });