+++ ChangeLog.md
@@ -1,5 +1,17 @@
 # Change Log - oav
 
+## 07/20/2022 3.0.4
+- GenerateExamples - Support data generation in byte format
+- ModelValidator - Support data validation in byte format
+- API Scenario
+  - Support readme.test.md run API test
+  - Fix step variable unresolved in newman collection
+  - Fix bugs about html report
+  - Aggregate reports into one per scenario file
+
+## 07/06/2022 3.0.3
+- Generate high quality examples from API Scenario tests
+
 ## 06/30/2022 3.0.2
 
 - traffic-converter